making constants value correct

+++ src/com/bsb/hike/chatHead/ChatHeadUtils.java
@@ -109,9 +109,9 @@ public class ChatHeadUtils
 
 	private static final String PARAMS = "params";
 
-	private static final long ONE_MONTH = (long)(30 * 24 * 60 * 60 * 1000);
+	private static final long ONE_MONTH = 30 * 24 * 60 * 60 * 1000l;
 
-	private static final long ONE_DAY = (long)(24 * 60 * 60 * 1000);
+	private static final long ONE_DAY = 24 * 60 * 60 * 1000l;
 
 	public static String msisdn;
